+++ .gitignore
@@ -144,7 +144,7 @@ dmypy.json
 # Cython debug symbols
 cython_debug/
 
-test.ipynb
+test*.ipynb
 test.py
 .vscode/settings.json
 input_files
